Fixed Legal Player Moves Check
Originally the player was able to cheat and make illegal moves (moves to any blank space) and would not allow legal attack moves (diagonal moves to capture enemy pieces). I have modified this code so it prompts for the row and column individually and then checks that the requsted move is available using the get_available_moves function, then returns the move if legal.